Cleanup Correctiondecoder
